+++ src/redux/helpers.js
@@ -1,4 +1,5 @@
 import { foldl, append } from 'funcadelic';
+import omitBy from 'lodash/omitBy';
 import { tagPaths } from '../constants/tagPaths';
 
 /**
@@ -34,7 +35,9 @@ export function mergeRelationships(existing, incoming) {
 export function mergeAttributes(existing, incoming) {
   if (!incoming) { return existing; }
 
-  return append(existing, incoming);
+  const omittedExisting = omitBy(existing, (_, key) => /^userDefinedField\d+$/.test(key));
+
+  return append(omittedExisting, incoming);
 }
 
 /**